Move Library folder to fix linker issue.
