+++ src/icm20948sensor.cpp
@@ -80,8 +80,6 @@ void ICM20948Sensor::save_bias(bool repeat) {
     // bool accel_set = bias_a[0] && bias_a[1] && bias_a[2];
     // bool gyro_set = bias_g[0] && bias_g[1] && bias_g[2];
     // bool mag_set = bias_m[0] && bias_m[1] && bias_m[2];
-
-    //Demo saved them to Preferences but that are part of the libarary we can't use for licising issues
     
     // if (accel_set) {
     //   // Save accel
